Moving Provider out of routing directory
