+++ graylog2-web-interface/src/components/common/Scratchpad.jsx
@@ -6,7 +6,7 @@ import ClipboardJS from 'clipboard';
 import teinte from 'theme/teinte';
 import { Alert, Button, MenuItem, SplitButton } from 'components/graylog';
 import { BootstrapModalConfirm } from 'components/bootstrap';
-import { ScratchpadContext } from 'routing/context/ScratchpadProvider';
+import { ScratchpadContext } from 'providers/ScratchpadProvider';
 /* NOTE: common components are cyclical dependencies, so they need to be directly imported */
 import InteractableModal from 'components/common/InteractableModal';
 import Icon from 'components/common/Icon';
