+++ graylog2-web-interface/src/providers/ScratchpadProvider.jsx
